fix: 修复yPosition类型不匹配问题
- 将 yPosition 初始化值改为 Float 类型
- 将 drawMarkdownContent 调用中的 MARGIN_TOP 参数转换为 Float

这修复了编译时的类型不匹配错误。

+++ app/src/main/java/com/childproduct/designassistant/util/PdfExporter.kt
@@ -62,7 +62,7 @@ object PdfExporter {
             drawTitle(canvas, paint, fileName)
             
             // 解析和绘制Markdown内容
-            var yPosition = MARGIN_TOP + 80
+            var yPosition = (MARGIN_TOP + 80).toFloat()
             yPosition = drawMarkdownContent(canvas, paint, markdownContent, yPosition)
 
             // 如果内容超过一页，创建新页面
@@ -71,7 +71,7 @@ object PdfExporter {
                 val newPageInfo = PdfDocument.PageInfo.Builder(PAGE_WIDTH, PAGE_HEIGHT, pdfDocument.pages.size + 1).create()
                 val newPage = pdfDocument.startPage(newPageInfo)
                 val newCanvas = newPage.canvas
-                yPosition = drawMarkdownContent(newCanvas, paint, markdownContent, MARGIN_TOP)
+                yPosition = drawMarkdownContent(newCanvas, paint, markdownContent, MARGIN_TOP.toFloat())
                 if (yPosition <= PAGE_HEIGHT - MARGIN_BOTTOM) {
                     page = newPage
                 } else {
